requirements-dev.txt
codespell
coverage
editorconfig-checker
flake8
isort
jupyter
mypy
pylint
sphinx
sphinx-autobuild
sphinx-click
sphinxcontrib-fulltoc
sphinxcontrib-napoleon
types-click
types-pytz
black
